feat(walker): add unified WalkContext and consolidate Option types (#217)
Refactor the walker package to provide a cleaner, more consistent API with
better contextual information for handlers.

## WalkContext

All handlers now receive a `*WalkContext` as their first parameter, providing
rich contextual information about the current node:

- JSONPath: Full JSON pointer path (e.g., `$.paths['/pets'].get`)
- PathTemplate: URL path template when in paths scope
- Method: HTTP method when in operation scope
- StatusCode: Status code when in response scope
- Name: Map key for named items (schemas, headers, etc.)
- IsComponent: True when walking components/definitions
- Context(): Access to context.Context for cancellation

Helper methods: InPathsScope(), InOperationScope(), InResponseScope()

## Unified Option Type

Consolidated `WalkInputOption` and `Option` into a single `Option` type:

Before:
  WalkWithOptions(
      WithFilePath("api.yaml"),
      WrapOption(WithSchemaHandler(fn)),  // Adapter needed
  )

After:
  WalkWithOptions(
      WithFilePath("api.yaml"),
      WithSchemaHandler(fn),  // No adapter needed
  )

## Test Reorganization

Split the monolithic walker_test.go into focused modules:
- walker_handlers_test.go: Handler-specific tests
- walker_oas2_test.go: OAS 2.0 specific tests
- walker_options_test.go: WalkWithOptions API tests
- walker_context_test.go: WalkContext feature tests

## Breaking Changes

Handler signatures changed from separate parameters to WalkContext:

Before:
  func(path string, method string, op *parser.Operation) Action

After:
  func(wc *WalkContext, op *parser.Operation) Action

Migration: Replace positional params with wc.JSONPath, wc.Method, etc.

Additional breaking changes:
- WalkInputOption type removed
- WrapOption() function removed
- WithMaxSchemaDepth() silently ignores invalid values

Closes #217
